Use constant instead of var for assembler cache flushes.
Citing Jerry Xia:

"The caches.c file used in boot and init has a function that does
"cache [var], offset($base)". This is fine for -O3 as the compiler
optimises [var] and treats it as a constant. However, -O0 does not,
and it is impossible to pass a variable in a 5-bit constant field.
A workaround is to hard-code [var] to a constant."