Remove use of system register in userspace context switch
kr1c was necessarily being used to point to the register stack frame.
Under CheriBSD this register is not available, so replace its use with
c26.  This reserves c26, the code fails to restore this user register
and corrupts it, however without leaking a secret (user's PCC).

Note that this aligns with the messages currently setting c26 to NULL,
effectively protecting against this corruption in the case of a
context switch due to sync message sending.