Remove knowledge of the namespace manager and memory manager from the kernel.
The namespace mgr is completely handled in user-space, by initializing the
frame appropriately before registration.

The memory manager/page-fault handler needs to be known by the kernel; this
is specified when an activation is registered.  In the default API that does
not specify this, the new activation inherits this from its creator.

A new system call allows setting the page-fault handler for an activation;
this allows bootstrapping.