Make progress on RISCV port
This adds in enough support to get a hello world in kernel. The kernel
now correctly links with the nanokernel. Basic physical book setup is
done in the nanokernel (split/merge), and get physical page are now
done. A few other placeholder functions have been hacked up to make this
work but still need to be done properly. I removed an unused argument
from the kernel main as it would have required passing a stack argument
on RISCV which would have been tedious.