Basic foundation implementation for RISCV
This adds a sha256 implementation and all foundation nanokernel functions (apart from the messaging ones).
Also adds a bit of extra glue to get to the secure load asm inside a secure loaded program.
Hashing is still untested, and likely will be until the system is working enough to perform the self test inside dedup.