Handle stack passing a little more generally
This still needs doing better. Currently this just always (tries) to
copy 4 caps across when changing stacks. It really requires entry stubs
to pass an integer of how much to copy (as entry stubs know the types of
the function). PLT stubs to do something similar to setbounds with a
correct length.